added get all appointemnts by doctor

+++ src/main/java/mycode/doctor_appointment_api/app/doctor/repository/DoctorRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface DoctorRepository extends JpaRepository<Doctor, Integer> {
 
     @EntityGraph(attributePaths = {"appointments"}, type = EntityGraph.EntityGraphType.FETCH)
-    Optional<Doctor> findById(Long id);
+    Optional<Doctor> findById(int id);
 
     @EntityGraph(attributePaths = {"appointments"}, type = EntityGraph.EntityGraphType.FETCH)
     Optional<Doctor> findByEmail(String email);
